Begining to implement the pattern library
after years of using SC i'm now understanding the difference between
patterns and streams! there shouldn't be any problems implementing _all_
of the patterns from `sclang`, though some work may need to be done at
the c-level. i'm going to look at the implementation in `sclang` for
sending pattern messages to a synth on the server (it's definitely just
a blob, but i'm not sure how or if supriya handles this yet).

- clean up (rather clean out) PatternCollector.py
  - replace with a `Bindings` module, that implements the different
    `sclang` bindings (Pmono, PmonoArctic, Pbind etc)
  - research the osc protocol for sending events to a registered synth
- work on docs (while the pattern library for sc users should be
  familiar, it will be horrifying to approach as a non sc user without
  any documentation)
- figure out a good way to deal with the shared key patterns (this will
  likely take some time and clever hacks, _bytecode hacks?! NOOOO...
  YESSS_)